LayoutTests/imported/w3c/web-platform-tests/html/semantics/embedded-content/the-object-element/object-in-object-fallback-2-expected.txt


FAIL Test <object> nesting inside <object> assert_equals: Should have loaded all should-load elements expected 12 but got 0
